Fix delete rented catalog

+++ frontend/src/components/Catalog.jsx
@@ -4,12 +4,16 @@ import {catalogAtom} from './HomePage'
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import {Link} from 'react-router-dom';
+import {customerAtom} from './Customers'
+
 const Catalog = () => {
     const {id} = useParams();
     const catalogId = parseInt(id);
     const [catalogs, setCatalogs] = useAtom(catalogAtom);
     const catalog = catalogs.find(cat => cat.id === catalogId);
     const navigate = useNavigate();
+
+    const [customers] = useAtom(customerAtom);
 
     const handleDelete = (e) => {
         e.preventDefault();
@@ -18,6 +22,16 @@ const Catalog = () => {
             const config = {
                 headers: { Authorization: `Bearer ${token}` }
             };
+
+            const isCatalogRented = customers.some(customer => {
+                return customer.rented.includes(catalogId);
+            });
+
+            if (isCatalogRented) {
+                window.alert('Cannot delete a catalog that is rented by a customer.');
+                return;
+            }
+
             const url = `http://localhost:3000/api/catalogs?id=${catalog.id}`;
             axios.delete(url, config)
             .then(response => {
